+++ lib/src/flutter/extensions.dart
@@ -15,3 +15,5 @@ export 'extensions/gesture_extension.dart';
 export 'extensions/padding_extension.dart';
 
 export 'extensions/context_extension.dart';
+
+export 'extensions/radius_extension.dart';